hospital/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-environ==0.12.0
environ==1.0
et_xmlfile==2.0.0
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
waitress==3.0.2
whitenoise==6.9.0
django-import-export==4.3.8
